requirements.txt
opencv-python
numpy
ultralytics
easyocr
transformers
torch
faiss-cpu
sentence-transformers
pyyaml
psutil
fuzzywuzzy
python-Levenshtein